# src/app/api/v1/contacts/route.ts
import { NextResponse, type NextRequest } from "next/server";
import { z } from "zod";
import { db } from "@/lib/db";
import { requireApiKey } from "@/lib/api-key-auth";
import { computeDedupeKey, findDuplicateContacts } from "@/lib/dedupe";

const createContactSchema = z.object({
  firstName: z.string().trim().min(1).max(200),
  lastName: z.string().trim().max(200).optional(),
  email: z.string().trim().toLowerCase().email().max(320).optional(),
  phone: z.string().trim().max(50).optional(),
  company: z.string().trim().max(200).optional(),
});

export async function GET(request: NextRequest) {
  const auth = await requireApiKey(request);
  if (auth.unauthorized) return auth.unauthorized;

  const contacts = await db.contact.findMany({
    where: { tenantId: auth.tenantId },
    orderBy: { createdAt: "desc" },
    take: 100,
  });
  return NextResponse.json({ contacts });
}

export async function POST(request: NextRequest) {
  const auth = await requireApiKey(request);
  if (auth.unauthorized) return auth.unauthorized;

  const body = await request.json().catch(() => null);
  const parsed = createContactSchema.safeParse(body);
  if (!parsed.success) {
    return NextResponse.json(
      { error: "Invalid input", details: parsed.error.flatten().fieldErrors },
      { status: 400 },
    );
  }

  const { tenantId } = auth;
  const duplicates = await findDuplicateContacts(tenantId, parsed.data);
  const contact = await db.contact.create({
    data: {
      tenantId,
      ...parsed.data,
      dedupeKey: computeDedupeKey(parsed.data),
    },
  });

  return NextResponse.json(
    { contact, possibleDuplicates: duplicates.map((d) => d.id) },
    { status: 201 },
  );
}